refactor: rename the Jest transformer files to camelCase
