+++ src/__tests__/jestTransformerBabelJest.test.ts
@@ -1,4 +1,4 @@
-import jestTransformerBabelJest from "../jest-transformer-babel-jest";
+import jestTransformerBabelJest from "../jestTransformerBabelJest";
 
 test("it exports a transformer object", () => {
 	expect(typeof jestTransformerBabelJest).toMatch("object");
