+++ src/jestTransformerBinaryFile.ts
